Finish basic functionality
Still isn't compiling, there is a bug in the linking.
