+++ Moves.hpp
@@ -1,0 +1,13 @@
+namespace moves
+{
+
+    enum moves 
+    { 
+        HIT,
+        STAND,
+        DOUBLE,
+        SPLIT,
+        SURRENDER 
+    };
+
+}
