+++ Shoe.hpp
@@ -1,52 +1,40 @@
 #ifndef SHOE_H
 #define SHOE_H
 
-#include <iostream>
-#include <vector>
-#include <string>
+#include <algorithm>
+#include <random>
+#include <chrono>
 
 using namespace std;
+
+string suit [13] = { "2", "3", "4", "5", "6", "7", "8", "9", "10", "10", "10", "10", "A" };
 
 class Shoe
 {
     private:
 
         vector<string> shoe;
-        string suit [13] = { "2", "3", "4", "5", "6", "7", "8", "9", "10", "10", "10", "10", "A" };
 
     public:
 
         Shoe(int decks) {
 
-            this->shoe.clear();
+            shoe.clear();
             int shoe_length = decks * 52;
 
             for (int i = 0; i < shoe_length; i++) {
-                this->shoe.push_back(this->suit[i % 13]);
+                shoe.push_back(suit[i % 13]);
             }
 
             unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
             std::default_random_engine e(seed);
-            shuffle(this->shoe.begin(), this->shoe.end(), e);
+            shuffle(shoe.begin(), shoe.end(), e);
 
         }
 
-        string draw() {
-            string card = this->shoe.back();
-            this->shoe.pop_back();
-            return card;
-        }
-
-        void print() {
-            for (int i = 0; i < shoe.size(); i++) {
-                cout << shoe[i] << " ";
-            }
-            cout << endl;
-        }
-
-        int get_shoe_size() {
-            return shoe.size();
-        }
+        int get_shoe_size();
+        string draw();
+        void print();
 };
 
 #endif
